@@ -2,21 +2,15 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace MediatrTests.Simple
+namespace MediatrTests.Simple;
+
+public class NotificationHandler1(IAssertService<NotificationMessage> assertService) : INotificationHandler<NotificationMessage>
 {
-    public class NotificationHandler1 : INotificationHandler<NotificationMessage>
+    private readonly IAssertService<NotificationMessage> assertService = assertService;
+
+    public Task Handle(NotificationMessage notification, CancellationToken cancellationToken)
     {
-        private readonly IAssertService<NotificationMessage> assertService;
-
-        public NotificationHandler1(IAssertService<NotificationMessage> assertService)
-        {
-            this.assertService = assertService;
-        }
-
-        public Task Handle(NotificationMessage notification, CancellationToken cancellationToken)
-        {
-            assertService.Process(notification);
-            return Task.CompletedTask;
-        }
+        assertService.Process(notification);
+        return Task.CompletedTask;
     }
 }
