@@ -1,7 +1,6 @@
 ﻿using MediatR;
 
-namespace MediatrTests.Simple
-{
-    public class Ping : IRequest<string>
-    { }
-}
+namespace MediatrTests.Simple;
+
+public class Ping : IRequest<string>
+{ }
