@@ -2,13 +2,12 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace MediatrTests.Simple
+namespace MediatrTests.Simple;
+
+public class PingHandler : IRequestHandler<Ping, string>
 {
-    public class PingHandler : IRequestHandler<Ping, string>
+    public Task<string> Handle(Ping request, CancellationToken cancellationToken)
     {
-        public Task<string> Handle(Ping request, CancellationToken cancellationToken)
-        {
-            return Task.FromResult("Pong");
-        }
+        return Task.FromResult("Pong");
     }
 }
